No change

+++ rib/hook_fns.py
@@ -208,17 +208,19 @@ def M_dash_and_Lambda_dash_pre_forward_hook_fn(
 
     with torch.inference_mode():
         in_dtype = in_grads.dtype
+        # Do we need float64 inside the einsum or afterwards?
+        # And can we go back to float32 after accumulation?
         M_dash = torch.einsum(
             einsum_pattern,
-            in_grads.to(torch.float64) / normalization_factor,
-            in_grads.to(torch.float64),
-        )
+            in_grads / normalization_factor,
+            in_grads,
+        ).to(torch.float64)
         # Concatenate the inputs over the hidden dimension
         in_acts = torch.cat(inputs, dim=-1)
         Lambda_dash = torch.einsum(
             einsum_pattern,
-            in_grads.to(torch.float64) / normalization_factor,
-            in_acts.to(torch.float64),
+            in_grads / normalization_factor,
+            in_acts,
         )
 
         _add_to_hooked_matrix(hooked_data, hook_name, data_key[0], M_dash)
